requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.9.0
markdown-it-py==3.0.0
mcp==1.3.0
mdurl==0.1.2
multidict==6.2.0
nest-asyncio==1.6.0
openai==1.72.0
pillow==11.2.1
playwright==1.51.0
propcache==0.3.1
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyee==12.1.1
Pygments==2.19.1
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.46.1
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.19.0
